requirements.txt
streamlit
python-telegram-bot
google-genai